+++ app/src/main/java/com/icam/interfaces/Classifier.java
@@ -9,5 +9,5 @@ import com.icam.models.Classification;
 public interface Classifier {
     String name();
 
-    Classification recognize(final int[] pixels);
+    Classification recognize(final float[] pixels);
 }
